feat(coupon-domain): extract coupon entities and repositories
into coupon-domain module
