+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/coupon/application/dto/CouponIssueResult.java
@@ -1,6 +1,7 @@
 package dev.be.coupon.api.coupon.application.dto;
 
-import dev.be.coupon.api.coupon.domain.IssuedCoupon;
+
+import dev.be.coupon.domain.coupon.IssuedCoupon;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
