+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/coupon/infrastructure/redis/config/RedisConfig.java
@@ -2,7 +2,7 @@ package dev.be.coupon.api.coupon.infrastructure.redis.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import dev.be.coupon.api.coupon.domain.Coupon;
+import dev.be.coupon.domain.coupon.Coupon;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
