+++ coupon/coupon-api/src/test/java/dev/be/coupon/api/coupon/application/CouponServiceTest.java
@@ -5,16 +5,15 @@ import dev.be.coupon.api.coupon.application.dto.CouponCreateResult;
 import dev.be.coupon.api.coupon.application.dto.CouponIssueCommand;
 import dev.be.coupon.api.coupon.application.dto.CouponIssueResult;
 import dev.be.coupon.api.coupon.application.exception.InvalidIssuedCouponException;
-import dev.be.coupon.api.coupon.domain.CouponRepository;
-import dev.be.coupon.api.coupon.domain.FakeUserRoleChecker;
-import dev.be.coupon.api.coupon.domain.IssuedCouponRepository;
-import dev.be.coupon.api.coupon.domain.exception.InvalidCouponTypeException;
-import dev.be.coupon.api.coupon.domain.exception.UnauthorizedAccessException;
 import dev.be.coupon.api.coupon.infrastructure.kafka.dto.CouponIssueMessage;
 import dev.be.coupon.api.coupon.infrastructure.kafka.producer.CouponIssueProducer;
 import dev.be.coupon.api.coupon.infrastructure.redis.AppliedUserRepository;
 import dev.be.coupon.api.coupon.infrastructure.redis.CouponCacheRepository;
 import dev.be.coupon.api.coupon.infrastructure.redis.CouponCountRedisRepository;
+import dev.be.coupon.domain.coupon.CouponRepository;
+import dev.be.coupon.domain.coupon.IssuedCouponRepository;
+import dev.be.coupon.domain.coupon.exception.InvalidCouponTypeException;
+import dev.be.coupon.domain.coupon.exception.UnauthorizedAccessException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.AfterEach;
