+++ coupon/coupon-domain/build.gradle
@@ -1,0 +1,11 @@
+jar.enabled = true
+
+dependencies {
+    implementation project(":support:common")
+    implementation 'org.springframework:spring-context'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+}
+
+tasks.withType(Test).configureEach {
+    useJUnitPlatform()
+}
