+++ coupon/coupon-domain/src/main/java/dev/be/coupon/domain/coupon/FailedIssuedCouponRepository.java
@@ -1,8 +1,13 @@
-package dev.be.coupon.kafka.consumer.domain;
+package dev.be.coupon.domain.coupon;
 
 import java.util.List;
 import java.util.UUID;
 
+/**
+ * coupon-api, coupon-kafka-consumer 양쪽에서 사용되는 발급 쿠폰 저장소입니다.
+ *
+ * 이를 위해 도메인 전용 모듈(coupon-domain)에서 추가했습니다.
+ */
 public interface FailedIssuedCouponRepository {
 
     FailedIssuedCoupon save(final FailedIssuedCoupon failedIssuedCoupon);
