+++ coupon/coupon-domain/src/main/java/dev/be/coupon/domain/coupon/IssuedCouponRepository.java
@@ -1,12 +1,11 @@
-package dev.be.coupon.api.coupon.domain;
+package dev.be.coupon.domain.coupon;
 
 import java.util.UUID;
 
 /**
  * coupon-api, coupon-kafka-consumer 양쪽에서 사용되는 발급 쿠폰 저장소입니다.
  *
- * 도메인 공유를 위한 현재 구조에서는 하나의 정의를 참조하고 있으며,
- * 추후 확장 시 도메인 전용 모듈(coupon-domain) 분리 및 의존성 정리가 필요할 수 있습니다.
+ * 이를 위해 도메인 전용 모듈(coupon-domain)에서 추가했습니다.
  */
 public interface IssuedCouponRepository {
